chore(routes): move non routes components out of the pages folder
