+++ lib/game/Waiting.jsx
