+++ pages/index.js
@@ -1,6 +1,6 @@
 import { StyleSheet, View } from "react-native";
 import React from "react";
-import Game from "./game/Game.jsx";
+import Game from "./tipsy/game.jsx";
 
 export default function App() {
     return (
