+++ pages/tipsy/game.jsx
@@ -1,9 +1,8 @@
 import React, { useState, useCallback, useEffect } from "react";
 import { Text, View, StyleSheet, TouchableOpacity } from "react-native-web";
-import GameStatus from "./GameStatus.jsx";
-import PropTypes from "prop-types";
-import gameApi from "../api/GameApi.jsx";
-import Cell from "./Cell.jsx";
+import GameStatus from "../../lib/game/GameStatus.jsx";
+import gameApi from "../../lib/game/api/GameApi.jsx";
+import Cell from "../../lib/game/Cell.jsx";
 
 const boardObstacles = [
     ["topleft", "exit", "top", "top", "top", "top", "topright"],
